backend/requirements0.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
apify-client>=1.6.0
openai>=1.12.0
requests>=2.31.0
